// src/api/generated/model/getRegulacoesCompletas200.zod.ts
export const getRegulacoesCompletas200DataItemIdRegExp = new RegExp(
  "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
);
export const getRegulacoesCompletas200DataItemCriticidadesRegulatoriasItemIdRegExp =
  new RegExp(
    "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
  );
export const getRegulacoesCompletas200DataItemCriticidadesRegulatoriasItemRegulacaoIdRegExp =
  new RegExp(
    "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
  );
export const getRegulacoesCompletas200DataItemCriticidadesRegulatoriasItemRegraQualidadeIdRegExp =
  new RegExp(
    "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
  );
export const getRegulacoesCompletas200DataItemCriticidadesRegulatoriasItemRegraQualidadeIdRegExpOne =
  new RegExp(
    "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
  );
export const getRegulacoesCompletas200DataItemCriticidadesRegulatoriasItemRegraQualidadeDimensaoIdRegExp =
  new RegExp(
    "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
  );

export const GetRegulacoesCompletas200 = zod.object({
  data: zod.array(
    zod.object({
      id: zod.uuid().regex(getRegulacoesCompletas200DataItemIdRegExp),
      epigrafe: zod.string(),
      orgao: zod.string(),
      descricao: zod.string(),
      dataInicio: zod.iso.datetime({ offset: true }),
      dataFim: zod.union([
        zod.iso.datetime({ offset: true }),
        zod.literal(null).nullable(),
      ]),
      criticidadesRegulatorias: zod
        .array(
          zod.object({
            id: zod
              .uuid()
              .regex(
                getRegulacoesCompletas200DataItemCriticidadesRegulatoriasItemIdRegExp,
              ),
            regulacaoId: zod
              .uuid()
              .regex(
                getRegulacoesCompletas200DataItemCriticidadesRegulatoriasItemRegulacaoIdRegExp,
              ),
            regraQualidadeId: zod
              .uuid()
              .regex(
                getRegulacoesCompletas200DataItemCriticidadesRegulatoriasItemRegraQualidadeIdRegExp,
              ),
            grauCriticidade: zod.string(),
            regraQualidade: zod.object({
              id: zod
                .uuid()
                .regex(
                  getRegulacoesCompletas200DataItemCriticidadesRegulatoriasItemRegraQualidadeIdRegExpOne,
                ),
              descricao: zod.string(),
              dimensao: zod.object({
                id: zod
                  .uuid()
                  .regex(
                    getRegulacoesCompletas200DataItemCriticidadesRegulatoriasItemRegraQualidadeDimensaoIdRegExp,
                  ),
                nome: zod.string(),
              }),
            }),
          }),
        )
        .optional(),
      createdAt: zod.union([
        zod.iso.datetime({ offset: true }),
        zod.literal(null).nullable(),
      ]),
      updatedAt: zod.union([
        zod.iso.datetime({ offset: true }),
        zod.literal(null).nullable(),
      ]),
    }),
  ),
});

export type GetRegulacoesCompletas200 = zod.input<
  typeof GetRegulacoesCompletas200
>;
export type GetRegulacoesCompletas200Output = zod.output<
  typeof GetRegulacoesCompletas200
>;